requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
